requirements.txt
Flask==3.0.3
wikipedia==1.4.0
python-dotenv==1.0.1
google-genai==0.3.0
requests==2.32.3
colorama==0.4.6
